order

+++ src/routes/(admin)/delete/+page.server.js
@@ -1,6 +1,6 @@
 import { deleteProduct } from "$lib/server/products";
-import { getProducts } from "$lib/server/products";
-
+importy { getProducts } from "$lib/server/products";
+import { requirePermission } from "$lib/server/permission"
 export async function load() {
   const menu = await getProducts();
   return { menu };
@@ -9,6 +9,7 @@ export const actions = {
   delete: async ({ request }) => {
     const data = await request.formData();
     const _id = data.get("_id");
+    await requirePermission(request.headers, { product: ["delete"] });
     await deleteProduct(_id);
     return { success: true };
   },
